Refactor GradientTiles example + bump typegpu version
